Added "projects" to code - Setup storage for "projectList" and "projectCounter" - projectList     - stores list of project names identified     by str generated from "task" plus the     projectCounter number at given time - projectCounter     - stores the current project projectCounter     number so that any new projects added after     reload will still have a unique name/number - added a project number to the "task" object
My thinking is that all the tasks will be stored
in one (big) array, but it will make the sorting
functions easier to run later in the project.

This is not ideal but without the benefits of
a linked database it'll be the most straightforward
to code.

HOPING run time will not be too slow.